dat-service-extension: renaming
