+++ data-protocols/ids/ids-core/README.md
@@ -1,0 +1,14 @@
+# IDS Core
+
+## Included Extensions
+
+- [data-protocols:ids:ids-spi](../ids-spi/README.md)
+
+## Configuration
+
+| KEY |  Description |
+|:---|:---|
+| edc.ids.daps.tokenurl | The token endpoint of the IDS DAPS |
+| edc.ids.daps.certificatepath | The file system path to the X.509 certificate |
+| edc.ids.daps.privatekeypath | The file system path to the signing private key |
+| edc.ids.daps.privatekeypassphrase | Optional private key passphrase |
