+++ data-protocols/ids/ids-core/src/main/java/org/eclipse/dataspaceconnector/ids/core/configuration/InvalidConfigurationException.java
@@ -1,0 +1,21 @@
+package org.eclipse.dataspaceconnector.ids.core.configuration;
+
+import org.eclipse.dataspaceconnector.spi.EdcException;
+
+public class InvalidConfigurationException extends EdcException {
+    public InvalidConfigurationException(String message) {
+        super(message);
+    }
+
+    public InvalidConfigurationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public InvalidConfigurationException(Throwable cause) {
+        super(cause);
+    }
+
+    public InvalidConfigurationException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
