--- data-protocols/ids/ids-daps-dat/src/main/java/org/eclipse/dataspaceconnector/ids/daps/Dat.java
@@ -1,31 +1,0 @@
-package org.eclipse.dataspaceconnector.ids.daps;
-
-import org.jetbrains.annotations.NotNull;
-
-import java.time.Instant;
-import java.util.Objects;
-
-public class Dat {
-    private final String token;
-    private final Instant expirationDate;
-
-    public Dat(@NotNull final String token, @NotNull final Instant expirationDate) {
-        Objects.requireNonNull(token);
-        Objects.requireNonNull(expirationDate);
-
-        this.token = token;
-        this.expirationDate = expirationDate;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public Instant getExpiration() {
-        return expirationDate;
-    }
-
-    public boolean isExpired() {
-        return expirationDate.isAfter(Instant.now());
-    }
-}
