--- data-protocols/ids/ids-daps-dat/src/main/java/org/eclipse/dataspaceconnector/ids/daps/DatService.java
@@ -1,6 +1,0 @@
-package org.eclipse.dataspaceconnector.ids.daps;
-
-public interface DatService {
-
-    Dat getDat() throws DatServiceException;
-}
