--- data-protocols/ids/ids-daps-dat/src/main/java/org/eclipse/dataspaceconnector/ids/daps/client/DapsIssuer.java
@@ -1,6 +1,0 @@
-package org.eclipse.dataspaceconnector.ids.daps.client;
-
-public interface DapsIssuer {
-
-    String getValue();
-}
