+++ data-protocols/ids/ids-identity-provider/src/main/java/org/eclipse/dataspaceconnector/ids/daps/DynamicAttributeTokenProvider.java
@@ -1,0 +1,6 @@
+package org.eclipse.dataspaceconnector.ids.daps;
+
+public interface DynamicAttributeTokenProvider {
+
+    DynamicAttributeToken getDynamicAttributeToken() throws DynamicAttributeTokenException;
+}
