+++ data-protocols/ids/ids-identity-provider/src/main/java/org/eclipse/dataspaceconnector/ids/daps/client/DapsClientException.java
