+++ data-protocols/ids/ids-identity-provider/src/main/java/org/eclipse/dataspaceconnector/ids/daps/client/IssuerProvider.java
@@ -1,0 +1,6 @@
+package org.eclipse.dataspaceconnector.ids.daps.client;
+
+public interface IssuerProvider {
+
+    String getIssuer();
+}
