+++ data-protocols/ids/ids-identity-provider/src/test/resources/test-certificate.crt
