+++ data-protocols/ids/ids-identity-provider/src/test/resources/test-rsa-encrypted-private-key.pem
@@ -1,0 +1,54 @@
+-----BEGIN ENCRYPTED PRIVATE KEY-----
+MIIJrTBXBgkqhkiG9w0BBQ0wSjApBgkqhkiG9w0BBQwwHAQIK2OeRjzplbICAggA
+MAwGCCqGSIb3DQIJBQAwHQYJYIZIAWUDBAEqBBDndPmCAW/BmqTXI9mo+f3zBIIJ
+UJNL/h8+3INN8Pw7B0o0RKrWB/UdlFfziwgwdjRDmb1H54KreEwctNjz+o/kRxo8
+PXulyZHcSYOjNVqaeT8WqSYHJKxw+uKyDyFJ1jXhZku8q+37eHyiM+HihFl4AJpc
+Qr+EaAaAD4ouHvxEu4S2SYm8PKxWiTBTuY4F2Kl369E2dUsjtW2F8JIJNUifGURZ
+bh8GjcFlSqdsZRGJA1O0RVLJT+jUi9p5pjOeo8cPGThVYwKQDaoqnjI8mHqafI1X
+ZD+V6s1mFA7OQ1II8ABIRXP4XA1tkVItqDNDI5mnmLrvR90YtCNGv8dJ/FX3m/ys
+qt0VXo5G6ObhtQ1we/hbYtmaYIizIZe+10kAPe8q7fDMCAFhoullZrtxlMZgw5kd
+4jFyRTnUNsgOYI83mjhIvRNegtaGXvmEkXg5o3Tba61kayfs/OA6a3w4ThDZboV1
+ZY5JADjgNJwHjuYZSvmPZ0OhjCU5hMsgnw60vsHs8Q9OCuEaszaxzVrVlvAr6F+e
+v2vAPQ8SH72rLW6VHWFxpOzd4+O7BMjA5kn0Sf3wiRTlv4oTUuVwv8mIHa6z4Fqp
+mxUARceVpbiwBdlY1OOi8SgC5/3Hwwr/w4/UkS+nJHstvCLRo+fVao4HrzpMXdYj
+e2XtVIseBlvTrQXA2XGdHPmB/zBqCnf2eHa7M9pG1RQzKrBCxHMGNhx+sFbO2ZJS
+kHCu3n4ccoDoEvMtVkB2Hfu8OoxPOILdw/BBHvoxxeEoT+sY4Hz+aH4BBrLg+nWE
+gwML5qOxhMh0XXUQdMb7/AlADxA51w5gh14Hkw+8DdfFwfKpDRGSkHMMdyq0dM88
+GYNDgyhL99pvIt2CZQcidFiORvRVOqbtHQcygfhOl8UC5UpC3fXxSAd6wt9l3LWk
+HzwnSg7qldhhiW2SciRM5Pe/nm+0f0Sd05MrKQuLQT8wyQccVs07j2shi3A2rZCW
+g6TS6YLrqgCK2Jg1RnI1N5dfBu63kSt5lDfHz3aOfir7HnamMBoqaNyTmBgCPiXC
+Lsh/lVyUm0Bxh4kQFbCdB0H593g+mNDTubRy1WchVDnahZKEHZ1ZRgXtehzr9J4/
+0Vh3FlUaMxC9MPvDbTYBsbXDfQJHYW7MhAiK4j7ihHp4kpDyOfTQlutQvGsShST+
+f9gUnnmzL6gwCTljmyiB0HE/LCZ6lNRsFKmD01s/q8ShVNmzIGlRYK95o3ABACZl
+CIRExKFLAdvhiy1w50TyqUAVJMEQ7MsKAolKYXR4RzDGwA87CFhCnISPdZUbkguK
+IgiPpINSqXZH8NRAPyo7JweNEYUwTDWBLUUX1OWObFC7cAfIRdZ5eKu3BrajcHT6
+jE6hZrd5HSLKJGT5I6OS/fKo1TqZhb+KcBR2XPH8io2gBuJ8e1StUeiED6BeuNOP
+1NeU4X5GtKx8fQzwqBTVqQDClVznfa+ZXW4CROR1Si0Oy6G/395cDL5eBQ6NdhNK
+ArNwdGrc0WtgRFYUtwGGH+Az8XEekL7kOD1UFJtKN0aJJEUEHXDG2KGBCxDgDNp/
+mbhbWmppogU3spPx0/T2Y6wxSKXWoS4rdzoeJ8A6Yj6i365me1WIm3E9dyrXJjgf
+ntKsKQeakozk966V+WRwrNjJq6RbawoGuHjUaiznU+GxNrjYAj3Z43/AdpVH9d6Z
+rjJ3WhDeZipo/QlHcYL70Q142wal/an1w8izrko+NzJHlVPSB33h1d9ZDUtac9z8
+V7xCt1R9AtBuF00BFDSQCDaP5BGWHeJkqfv2UB47+mBMRkO5KxYL5v5s1pPHZuwt
+fHO5aZhiBFemSBz7bGvDdkK4EE0Wok1bI+Zw7D51ALEstFKyO9F8Zx0vPgK4TSlo
+I6MxbDi9sf1czVB163qZ/zFPnfR/AIRU+5Xizi0yJkXUhKS6KRRSWag9e4Qj5WrA
+NjJH96LhyQB0tHcVH2o5pbeuZPkod6ZajnHX5k8Cc0IwJan6ilsI5D0HqfvOzQfY
+8ucA8tPyK86oCLi7tXi3qGtOWwagNlaBxgLzmVFpCRxbvTPALgGTA6C5mC29jsO1
+hisb5OdZiIp9vqZf0os+SdIsUXe4OLvToKjmskDuKGfEsh5DzcupfUOpbfw9ayYT
+pxvCIiRwG19Rx1GydWGznVs8/LEUrqqf02TBPT3KNhU9MV8RNG3bGujSr4Crj6nx
+agBCjLsPOVHpvbffzHJasyBXNbYdtSeXFHtugztz6Gm/Ecx4FNlIWHr4pKw6u42n
+dT3q1ku15IiuSNb6prfRGqeSbvyt3Fq7tnIRd+NTVhAxJn8UPOoEPIQ3yByrDLpt
+YOpwuqMKSduX6vYhmY4h+zt2JG2yg6eI11vkKxzpVPmdlbS48T1VHKlwUh3O0/jT
+w/HaArutpDx0sftF3HvG+84SahTvtmqOvmDDhBJsuKJqy62w262mmMb76Ned5rbK
+BYWW95RaGDuzbeu/tDt71mbz6dIPSNNjVw3vS0WOY+inIxAzoUx7i7FtK9T+LKqC
++8YcQrBik/wsQvuXOx2U4HoZ5u3L86DxH8UNwvfVH5h2wxUCiBvlg3zeY3MP4iRX
+wQ/vPLrbRiu66KMJCMzmE0PfPrbHSOUNloH3IOyufmqeiwavRk3HMZbr28+WZZbX
+czWOmwOz250Jdr0WC52gYw6gI5JrMhjtyHvagBjcDqCe8bakZBRuA6iOvlT4xSOq
+XugszvCvQ/tcY/64+Op5yztzxIhO57I9pgp4Cv1abMK8HcFN8ty7m2eagrYrI1Gv
+Amegbp0e5qVNgUO6mJKIR8GoRFIOX8rNVKHsJ4HygsUJN0ulNU2qPL7jJRGW2yR5
+0vDsgCBv0t6lyIsjftDGORDonZ2cEhAVRu5XmHRhaMhnc7kiW5DG8sCuEXCMC83z
+w1NMdo0kymkqaOObOOz8T3Ehxx991TNsImpTkO4VkNyaqYPRTyaWGAMSKtNHNdz1
+ijB1d65r6IrT3in12RwUqDWUSmW1qbg082BMPT3vFSYli1D8vZKFX0/gIqgvg5rl
+o8L6OWKahOrRLUKH5BhdXEy2uEIOsFYozgUJ1DSB5uFT6MZfs/a0IOwcRHSJUGny
+hq0UyzO3igbEJVuRF8u1zyGbowSuQqLrXPSc9hxYLPm3vYwxDrZ0iSRbYUjQXjUz
+rTGP7e8Vvd5xzJJipLv8c+1YQ/NkxKY84ZqtA4HmnUgH
+-----END ENCRYPTED PRIVATE KEY-----
