+++ data-protocols/ids/ids-spi/src/main/java/org/eclipse/dataspaceconnector/ids/spi/configuration/ConfigurationProvider.java
@@ -1,0 +1,14 @@
+package org.eclipse.dataspaceconnector.ids.spi.configuration;
+
+import java.net.URI;
+import java.net.URL;
+import java.nio.file.Path;
+import java.util.Optional;
+
+public interface ConfigurationProvider {
+    Path resolveCertificatePath();
+    Path resolveKeyPairPath();
+    Optional<String> resolveKeyPairPassphrase();
+
+    URL resolveDynamicAttributeProvisionServiceTokenUrl();
+}
